src/backend/api/requirements.txt
Flask==1.1.2
mongoengine==0.22.0
PyJWT==1.7.1
bcrypt==3.2.0
Werkzeug==1.0.1
